Move utils/rules to search as they are only used there
